+++ packages/yourloops/components/switch-role/index.tsx
@@ -30,12 +30,14 @@ import React from "react";
 import { useTranslation } from "react-i18next";
 import bows from "bows";
 
+import { HcpProfession } from "../../models/hcp-profession";
+import { SwitchRoleDialogsProps, SwitchRoleToHcpSteps } from "./models";
 import metrics from "../../lib/metrics";
 import { useAuth } from "../../lib/auth";
 import { useAlert } from "../utils/snackbar";
-import { SwitchRoleDialogsProps, SwitchRoleToHcpSteps } from "./models";
 import SwitchRoleConsequencesDialog from "./consequences-dialog";
 import SwitchRoleConsentDialog from "./consent-dialog";
+import SwitchRoleProfessionDialog from "./profession-dialog";
 
 const log = bows("SwitchRoleDialogs");
 
@@ -44,35 +46,30 @@ function SwitchRoleDialogs(props: SwitchRoleDialogsProps): JSX.Element {
   const { switchRoleToHCP, user } = useAuth();
   const alert = useAlert();
   const [switchRoleStep, setSwitchRoleStep] = React.useState<SwitchRoleToHcpSteps>(SwitchRoleToHcpSteps.none);
+  const [feedbackConsent, setFeedbackConsent] = React.useState<boolean>(false);
 
   if (user === null) {
     throw new Error("User must be looged-in");
   }
 
   const role = user.role;
-  const handleSwitchRoleToConditions = (accept: boolean): void => {
-    const nextStep = accept ? SwitchRoleToHcpSteps.consent : SwitchRoleToHcpSteps.none;
-    setSwitchRoleStep(nextStep);
-    if (!accept) {
-      props.onCancel();
+  const handleSwitchRoleToConditions = (): void => {
+    setSwitchRoleStep(SwitchRoleToHcpSteps.consent);
+  };
+
+  const handleSwitchRoleToUpdate = async (hcpProfession: HcpProfession): Promise<void> => {
+    try {
+      await switchRoleToHCP(feedbackConsent, hcpProfession);
+      metrics.send("switch_account", "accept_terms");
+    } catch (reason: unknown) {
+      alert.error(t("modal-switch-hcp-failure"));
+      log.error("switchRoleToHCP", reason);
     }
   };
-  const handleSwitchRoleToUpdate = (accept: boolean, feedbackConsent: boolean): void => {
-    const nextStep = accept ? SwitchRoleToHcpSteps.update : SwitchRoleToHcpSteps.none;
-    setSwitchRoleStep(nextStep);
 
-    if (accept) {
-      switchRoleToHCP(feedbackConsent)
-        .then(() => {
-          metrics.send("switch_account", "accept_terms");
-        })
-        .catch((reason: unknown) => {
-          alert.error(t("modal-switch-hcp-failure"));
-          log.error("switchRoleToHCP", reason);
-        });
-    } else {
-      props.onCancel();
-    }
+  const handleSwitchRoleToProfession = (feedback: boolean): void => {
+    setFeedbackConsent(feedback);
+    setSwitchRoleStep(SwitchRoleToHcpSteps.profession);
   };
 
   React.useEffect(() => {
@@ -88,9 +85,19 @@ function SwitchRoleDialogs(props: SwitchRoleDialogsProps): JSX.Element {
       <SwitchRoleConsequencesDialog
         title="modal-switch-hcp-title"
         open={switchRoleStep === SwitchRoleToHcpSteps.consequences}
-        onResult={handleSwitchRoleToConditions}
+        onAccept={handleSwitchRoleToConditions}
+        onCancel={props.onCancel}
       />
-      <SwitchRoleConsentDialog open={switchRoleStep === SwitchRoleToHcpSteps.consent} onResult={handleSwitchRoleToUpdate} />
+      <SwitchRoleConsentDialog
+        open={switchRoleStep === SwitchRoleToHcpSteps.consent}
+        onAccept={handleSwitchRoleToProfession}
+        onCancel={props.onCancel}
+      />
+      <SwitchRoleProfessionDialog
+        open={switchRoleStep === SwitchRoleToHcpSteps.profession}
+        onAccept={handleSwitchRoleToUpdate}
+        onCancel={props.onCancel}
+      />
     </React.Fragment>
   );
 }